moved logic into configuration controller from the system controller and added byname bydatatypes and bytypes routes
